requirements.txt
# HTTP requests library
requests==2.31.0

# Type checking and data validation
typing-extensions==4.9.0
dataclasses==0.6
pydantic==2.5.3

# Environment variable management
python-dotenv==1.0.1

# Date and time handling
pytz==2024.1

# Optional but recommended packages
# Async HTTP requests for better performance
aiohttp==3.9.3

# For handling rate limiting and retries
ratelimit==2.2.1
backoff==2.2.1

# For logging and monitoring
structlog==24.1.0

# For testing
pytest==8.0.0
pytest-asyncio==0.23.5
responses==0.24.1

# For development
black==24.1.1
flake8==7.0.0
mypy==1.8.0